+++ Runtime/AssetFolderAttribute.cs
@@ -3,7 +3,7 @@ using System.Diagnostics;
 namespace SaintsField
 {
     [Conditional("UNITY_EDITOR")]
-    [System.AttributeUsage(System.AttributeTargets.Field | System.AttributeTargets.Property, Inherited = true, AllowMultiple = false)]
+    [System.AttributeUsage(System.AttributeTargets.Field | System.AttributeTargets.Property)]
     public class AssetFolderAttribute: FolderAttribute
     {
         public AssetFolderAttribute(string folder="Assets", string title="Choose a folder inside assets", string groupBy = "") : base(folder, title, groupBy)