+++ Runtime/TreeDropdownAttribute.cs
@@ -1,14 +1,10 @@
-using System;
-using System.Collections.Generic;
 using System.Diagnostics;
-using SaintsField.Interfaces;
-using SaintsField.Utils;
-using UnityEngine;
 
 // ReSharper disable once CheckNamespace
 namespace SaintsField
 {
     [Conditional("UNITY_EDITOR")]
+    [System.AttributeUsage(System.AttributeTargets.Field | System.AttributeTargets.Property | System.AttributeTargets.Method | System.AttributeTargets.Parameter)]
     public class TreeDropdownAttribute: PathedDropdownAttribute
     {
         public TreeDropdownAttribute(string funcName = null, EUnique unique = EUnique.None): base(funcName, unique)